+++ torchmetrics/functional/__init__.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from torchmetrics.functional.audio.pit import pit, pit_permutate
-from torchmetrics.functional.audio.sdr import sdr
+from torchmetrics.functional.audio.sdr import scale_invariant_signal_distortion_ratio, sdr, signal_distortion_ratio
 from torchmetrics.functional.audio.si_sdr import si_sdr
 from torchmetrics.functional.audio.si_snr import si_snr
 from torchmetrics.functional.audio.snr import scale_invariant_signal_noise_ratio, signal_noise_ratio, snr
@@ -129,7 +129,9 @@ __all__ = [
     "rouge_score",
     "sacre_bleu_score",
     "sdr",
+    "signal_distortion_ratio",
     "si_sdr",
+    "scale_invariant_signal_distortion_ratio",
     "si_snr",
     "scale_invariant_signal_noise_ratio",
     "snr",